Add substring searching support
This now makes flatcurve RFC 3501 compliant, at the expense of much
larger database sizes (since we have to manually store each substring
ourselves, as Xapian does not have built-in support).

For now, we do need to make this the default though, because of the
compliancy factor. Provide an option for a user to disable substring
seearching (reverting back to the previous behavior) instead.