Distinguish between full compaction and standard compaction
Standard compaction is recommended if it is expected that DB will grow
further. An auto triggered compaction should assume this.

Use the FULL compaction when a user explicitly triggers and
optimization. This makes adding new messages more expensive, but results
in much smaller DB sizes (which is probably what the user expects when
running the optimize command).