Add support for rotating write DB
Xapian loses performance when its DB gets larger, so we want to keep the
"current" (write target) DB to a certain size. If we hit the size
threshold, move the DB to a "old" (read-only) target DB and start
writing to a new writable DB.

Still missing: need to add code to delete messages from the "old" DBs
(that will be the only allowable write action to those DBs).