Use DB_NO_SYNC for Xapian::WritableDatabase
This setting means there is a small chance that indexed data will not be
properly committed to the DB (e.g. power failure), but that's a fair
trade for the vastly reduced storage I/O.

Worst-case scenario, Dovecot would have to re-index some messages the
next time the mailbox is accessed - there is no data loss as last UID
will catch this scenario.