Improve logic to determine when to rotate current index
When we should rotate the "current" index depends on a variety of factors
local to the current system. Thus, the algorithm to decide when to rotate
should not do so based on data size thresholds - it should do so based on
real-world performance values. In other words: if committing to the DB is
taking too long, then rotate the index.

Rotation only occurs when the DB is closed. This allows rotate_size to
control when doing large batch updates (e.g., indexing a large mailbox).
Time-based rotation is intended to capture the single updates (e.g., move to
Trash) that take a long time.

Added a configuration option to define this value (default: 5 seconds).