Skip indexing if the document already exists
If the document exists, this means that we have already indexed that
message. This could happen if two processes are updating the DB at the
same time.

This is unfortunate FTS API design: the last UID is checked first. If lower
than mailbox size, API triggers reindex. However, last UID may be
incorrect because other indexing process may not yet have finished.
The current indexing process will be locked until that process is
complete - the last UID may have changed by this time, but there is no
way to inform the API that there is no indexing to be done.