Fix assert-crash due to excessive scores length
For searches that hit multiple mailboxes (virtual mailboxes is an easy way
to force a multisearch), an assert-crash will happen if a message
matches both the main query and internal "maybe" queries.

Fix by ensuring that scores are only added once per UID.

This was broken in single mailbox search as well, but it was ignored there
as there was no assert check on scores length (the extra entries were ignored
during processing).

In order to calcuate scores correctly (now that we are only adding them once),
refactor result iteration to only run the maybe query once for all maybe
queries.  Add the main query to this query for score calculation purposes.
This is a bit of an optimization, so nice.

Fixes GitHub Issue #61